Upgrade pytest dependency to 7.0.1.
I get an error like this when I try to run the tests:

TypeError: required field "lineno" missing from alias

This was apparently fixed in a later version than we specify in
requirements.txt: https://github.com/pytest-dev/pytest/discussions/9195

This commit moves us from 5.4.3, which is 3 years old, to 7.0.1. We can't move
to the latest version, 7.4.0, because it only supports Python >= 3.7, and we're
stuck on 3.6, for now.