Make more improvements.
- Merge the functionality of _preprocess_req into _prepare_request. These are
always used together, so it doesn't really make sense to have them as separate
functions. Additionally, have _prepare_request do everything needed to transform
the request data for transmission. Namely, encode the request string as UTF-8
bytes. Finally, make _prepare_request a method of Notecard. There's no reason
for it to be a free function.
- Refactor low-level serial comms by adding a new function, _transact. This
replaces _transmit and _transmit_and_receive. It acquires the serial lock and
uses the transaction manager. It will read a response from the Notecard if the
rsp_expected parameter is True.