Fix a couple problems.
- Fail example scripts if an exception is raised at any point.
- Rework `_crc_add` to produce a reliable CRC. This requires us to rely more on
string manipulation and less on python dicts. Prior to this commit, the CRC was
computed over the string representation of the request (via `json.dumps`) and
then it was added to the request dict. Then, when it was time to serialize the
request for transmission and we converted the dict to a string again, because
the CRC field had been added, the order of the fields changed and the CRC was
invalidated. This is because Python dicts are fundamentally unordered
collections.
- Add missing timeout reset to `OpenSerial`'s `receive` method.
- Increase cpy_example.py UART RX buffer size. I was seeing the end of responses
get cut off with the default size (64). Increased to 128.