fix: Delay for 50 ms between I2C data queries.
This is something I neglected to add when aligning the I2C code with what's in
note-c. The result is that we'd hammer the Notecard, asking if there was data
available. This actually worked with some hosts, but on others, the querying
overwhelmed the Notecard, and the card would actually get into an error state
(solid red LED).