When checking CRC, treat sequence number and CRC as integers.
Prior to this commit, the hex strings for the CRC and sequence number from the
Notecard were compared against hex strings computed by the host. If the casing
of the hex digits differs between host and Notecard, there can be false CRC
errors. Treating all values as integers gets rid of this problem.